Move stoppable to core, implement Loop with stoppable
