+++ hybrid/core.py
@@ -13,10 +13,11 @@
 # limitations under the License.
 
 import os
-from collections import namedtuple, defaultdict
-from copy import deepcopy
 import operator
 import logging
+import threading
+from collections import namedtuple, defaultdict
+from copy import deepcopy
 
 from plucky import merge
 import dimod
@@ -439,6 +440,60 @@ class Runnable(traits.StateTraits):
         """Composition of runnable components (L-to-R) returns a new
         runnable Branch."""
         return Branch(components=(self, other))
+
+
+def stoppable(cls):
+    """Extends a `Runnable` subclass with a `stop_signal` semaphore/event, and
+    amends the existing `halt` and `next` methods to signal stop via the
+    semaphore, and reset the stop signal on run completion, respectively.
+
+    Example:
+        A Runnable block that accepts `timeout` and on `run` blocks for up to
+        `timeout` seconds. It can be interrupted via call to `stop`, in which
+        case blocks shorter than the timeout interval::
+
+            @stoppable
+            class StoppableSleeper(Runnable):
+                def next(self, state, timeout=None, **runopts):
+                    self.stop_signal.wait(timeout=timeout)
+                    return state
+
+            >>> sleeper = StoppableSleeper(timeout=30)
+            >>> sleeper.run(state)
+            <Future at 0x7fbb575e6f60 state=running>
+
+            >>> sleeper.stop()
+            >>> sleeper.timers
+            <snipped>
+                'dispatch.next': [13.224211193970405]
+
+    """
+    if not issubclass(cls, Runnable):
+        raise TypeError("Runnable subclass expected as the decorated class")
+
+    orig_init = cls.__init__
+    orig_halt = cls.halt
+    orig_next = cls.next
+
+    def __init__(self, *args, **kwargs):
+        orig_init(self, *args, **kwargs)
+        self.stop_signal = threading.Event()
+
+    def halt(self):
+        result = orig_halt(self)
+        self.stop_signal.set()
+        return result
+
+    def next(self, state, **runopts):
+        result = orig_next(self, state, **runopts)
+        self.stop_signal.clear()
+        return result
+
+    cls.__init__ = __init__
+    cls.halt = halt
+    cls.next = next
+
+    return cls
 
 
 class HybridSampler(dimod.Sampler):
